common : separate whisper sources
