+++ examples/talk-llama/talk-llama.cpp
@@ -3,6 +3,7 @@
 
 #include "common-sdl.h"
 #include "common.h"
+#include "common-whisper.h"
 #include "whisper.h"
 #include "llama.h"
